Restructure app.
