+++ src/assets/css/calculator.css
