+++ src/components/childComponents/quotes/RenderQuotes.js
